Store authorized guild IDs in the user token
This is a little difficult to follow as we want to update the token immediately if we request the list of guilds from Discord for the front-end and discover it has changed. It's important to do that as otherwise the user won't have immediate access to resources that check this when they join a new guild, but we don't want to query Discord for every single request to the API when we're loading graph data.

It would be much easier if we could replace the token from any request (it's likely technically possible as it is a cookie, but there is no API exposed by next-auth), and we work around that by indicating to the client it needs to issue a session update to next-auth, which we then look for and refresh the in-token guild list.

So we're not trusting the client, we also refresh the guild list in the token server-side if it is more than 10 minutes old. That would be the maximum time a user could access data from a guild they've been removed from if they didn't trigger, or were blocking, the client-side validation.